Set "backend" from __init__ to avoid the dreaded "global".
